requirements.txt
PyMuPDF
pytesseract
pandas
streamlit>=1.38.0
docxtpl
google-genai
Pillow
python-dotenv
openpyxl
playwright
pytest-playwright
